requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cssselect2==0.4.1
Django==3.2.10
django-ckeditor==6.2.0
django-js-asset==1.2.2
easy-thumbnails==2.8
lxml==4.6.4
Pillow==8.4.0
pytz==2021.3
reportlab==3.6.3
sqlparse==0.4.2
svglib==1.1.0
tinycss2==1.1.1
webencodings==0.5.1
